Add support for 32-bit VM numeric types via feature flag
Introduces a `number-32` feature flag to optionally switch the VM, compiler,
and host interop from 64-bit to 32-bit numeric types (`i32`, `u32`, `f32`).
Refactors core representations, modules, and APIs to use canonical numeric
type aliases, ensuring consistent type handling throughout the codebase.
Updates modules, tests, and serialization logic to account for 32-bit and
64-bit configurations, improving portability and enabling reduced memory
footprints on constrained targets.

+++ crates/rune/src/runtime/const_value.rs
@@ -308,10 +308,10 @@ impl ConstValue {
     /// ```
     pub fn as_integer<T>(&self) -> Result<T, RuntimeError>
     where
-        T: TryFrom<i64> + TryFrom<u64>,
+        T: TryFrom<i128> + TryFrom<u128>,
     {
         match self.kind {
-            ConstValueKind::Inline(Inline::Signed(value)) => match value.try_into() {
+            ConstValueKind::Inline(Inline::Signed(value)) => match i128::from(value).try_into() {
                 Ok(number) => Ok(number),
                 Err(..) => Err(RuntimeError::new(
                     VmErrorKind::ValueToIntegerCoercionError {
@@ -320,7 +320,7 @@ impl ConstValue {
                     },
                 )),
             },
-            ConstValueKind::Inline(Inline::Unsigned(value)) => match value.try_into() {
+            ConstValueKind::Inline(Inline::Unsigned(value)) => match u128::from(value).try_into() {
                 Ok(number) => Ok(number),
                 Err(..) => Err(RuntimeError::new(
                     VmErrorKind::ValueToIntegerCoercionError {
